Write XML file instead of printing to stdout
I was going to send the XML to stdout and use redirection to write the
output to a file, but I realized that the info prompts are also going to
stdout, and they get redirected, too.

So now the info prompts go to stdout and the xml is writted to a file
`_.xml` in the same folder as the MP3 files.